trading_engine/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
motor>=3.1.1
pymongo>=4.3.3
python-dotenv>=1.0.0
pydantic>=1.10.7
aiohttp>=3.8.4
backoff>=2.2.1
base58>=2.1.0